requirements.txt
psycopg2_binary==2.9.3
tweepy==4.8.0